requirements.txt
# Core LangChain 0.3 - Latest versions as of December 2024
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0  # For OpenRouter integration

# OpenAI SDK for OpenRouter compatibility
openai>=1.0.0
tiktoken>=0.5.0  # For token counting

# Pydantic 2 (required for LangChain 0.3)
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0  # Excel file handling
numpy>=1.24.0

# Configuration and Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Date/Time handling
python-dateutil>=2.8.2
holidays>=0.35  # Holiday detection for temporal analysis

# Geospatial Analysis (for tower location analysis)
geopy>=2.4.0
haversine>=2.8.0

# Visualization (optional, for future enhancements)
matplotlib>=3.7.0
plotly>=5.18.0

# CLI and Progress
click>=8.1.0
rich>=13.7.0  # For formatted console output
tqdm>=4.66.0  # Progress bars

# Type hints and validation
typing-extensions>=4.8.0

# Network analysis
networkx>=3.2.0

# HTTP client for API calls
httpx>=0.25.0
aiohttp>=3.9.0

# Logging and monitoring
loguru>=0.7.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality (optional)
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0